Whew!
Version 2.5 is out.  A few changes are out, and the whole thing is commented.  It took forever, and I found a few problems with the code just by looking at each line.  Do this for every file people, so that when you look back at it in a year, you will not think, "What was I doing???"